Check the WSU Alerts feed every minute and display when active
Check the feed used for WSU alerts and display the most recent
item in the emergency category. If no items are available, then
do not show an alert.

The feed is cached for a minute and we cache the entire HTML block
for a minute. In the future, we'll do this smarter.