Initial implementation of the WSU Home overlay feature
Currently accounts for takeover overlays. If the option is chosen
in the side area meta box, then Javascript is loaded that applies
the `noscroll` class to `body`. It then looks for a section that
has been wrapped in an `overlay-wrapper` class and fixes that to
fill the document.

Upon dismissal, a cookie is set that expires in 7 days so that
repeat visitors do not see the overlay after initial dismissal.